[REFACTOR/#111] 상품 목록 반환 Mapper 분리
