+++ app/src/main/java/com/napzak/market/data/product/mapper/MarketProductBuyItemsMapper.kt
@@ -1,0 +1,12 @@
+package com.napzak.market.data.product.mapper
+
+import com.napzak.market.data.product.dto.request.MarketProductBuyItemsRequest
+import com.napzak.market.domain.explore.model.ProductListFilter
+
+fun ProductListFilter.toMarketProductBuyRequest(): MarketProductBuyItemsRequest =
+    MarketProductBuyItemsRequest(
+        storeId = this.storeId,
+        sortOption = this.sortOption,
+        isOnSale = this.isOnSale,
+        genreIds = this.genreId,
+    )
