+++ app/src/main/java/com/napzak/market/data/product/mapper/MarketProductSellItemsMapper.kt
@@ -1,13 +1,13 @@
 package com.napzak.market.data.product.mapper
 
-import com.napzak.market.data.product.dto.request.ProductSellItemsRequest
+import com.napzak.market.data.product.dto.request.MarketProductSellItemsRequest
 import com.napzak.market.domain.explore.model.ProductListFilter
 
-fun ProductListFilter.toProductSellRequest(): ProductSellItemsRequest =
-    ProductSellItemsRequest(
+fun ProductListFilter.toMarketProductSellRequest(): MarketProductSellItemsRequest =
+    MarketProductSellItemsRequest(
+        storeId = this.storeId,
         sortOption = this.sortOption,
         isOnSale = this.isOnSale,
         isUnopened = this.isUnopened,
         genreIds = this.genreId,
-        nextCursor = "",
     )
